Next.js Routing: bug fix and add language setting
