+++ src/app/[lang]/layout.tsx
@@ -1,26 +1,14 @@
-import type { Metadata } from 'next';
-import './globals.css';
-import { NavLinks } from './ui/nav-links';
+type LangLayoutProps = Readonly<{
+  children: React.ReactNode;
+}>;
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+const LangLayout: React.FC<LangLayoutProps> = ({ children }) => {
+  return (
+    <>
+      {children}
+      <div style={{ position: 'absolute', bottom: '0' }}>Lang Layout!!</div>
+    </>
+  );
 };
 
-export default function RootLayout({
-  children,
-  auth,
-}: Readonly<{
-  children: React.ReactNode;
-  auth: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body>
-        <NavLinks />
-        <div>{auth}</div>
-        <main>{children}</main>
-      </body>
-    </html>
-  );
-}
+export default LangLayout;
