+++ src/app/dictionaries.ts
@@ -2,8 +2,8 @@ import 'server-only';
 
 // NOTE(hajae): 참고 https://nextjs-ko.org/docs/app/building-your-application/routing/internationalization
 const dictionaries = {
-  en: () => import('./dictionaries/en.json').then((module) => module.default),
-  ko: () => import('./dictionaries/ko.json').then((module) => module.default),
+  en: () => import('./[lang]/dictionaries/en.json').then((module) => module.default),
+  ko: () => import('./[lang]/dictionaries/ko.json').then((module) => module.default),
 } as const;
 
 export const getDictionary = async (locale: keyof typeof dictionaries) => {
